# REST/swagger.yaml
swagger: "2.0"
info:
  description: "This is book store server, you able send request to get list of books, insert book, delete book."
  version: "1.0.0"
  title: "Bookstore API"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
- name: "book"
  description: "Everything about your Pets"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "https"
- "http"
paths:
  /books:
    get:
      tags:
      - "book"
      summary: "Get a list of all the books"
      description: ""
      operationId: "addPet"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
  /book:
    post:
      tags:
      - "book"
      summary: "Add a book to list"
      description: ""
      operationId: "updateBook"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Book object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Book"
      responses:
        "200":
          description: "Book is added to the database"
  /book/{bookId}:
    get:
      tags:
      - "book"
      summary: "Get specific book by bookId"
      description: ""
      operationId: "getBook"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "bookId"
        description: "Get specific book by bookId"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
              $ref: "#/definitions/Book"
    delete:
      tags:
      - "book"
      summary: "Delete book in list"
      description: ""
      operationId: "deleteBook"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "path"
        name: "bookId"
        description: "Remove book in store by bookId"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Book is deleted"
definitions:
 
  Book:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      author:
        type: "string"
    xml:
      name: "Book"
  
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"